+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 0.1.475
 
+* add APPLY to FT.AGGREGATE
+* fix FT.CREATE "TAG" refence
+
+# 0.1.475
+
 * add module load/unload commands
 * add search test harness and search / suggest tests